Added correct current pace on top of the average pace. Current pace is split into small windows of time, have to fix this and reflect the change in the animation. Then, find a way to calculate the optimal pace based on user input

+++ lib/widgets/elapsed_time_provider.dart
@@ -1,12 +1,11 @@
-//#pace
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final elapsedTimeProvider = StateProvider<String>((ref) => '00:00:00');
 
-final elapsedTimeProviderInSeconds = Provider<int>((ref) {
+final elapsedTimeProviderInSeconds = Provider<double>((ref) {
   final timeString = ref.watch(elapsedTimeProvider);
   final parts = timeString.split(':');
-  return int.parse(parts[0]) * 3600 +
-      int.parse(parts[1]) * 60 +
-      int.parse(parts[2]);
+  return double.parse(parts[0]) * 3600 +
+      double.parse(parts[1]) * 60 +
+      double.parse(parts[2]);
 });
